requirements.txt
asgiref==3.5.0
autopep8==1.6.0
backports.zoneinfo==0.2.1
colorama==0.4.4
Django==4.0.4
djangorestframework==3.13.1
Jinja2==3.1.1
MarkupSafe==2.1.1
pdfkit==1.0.0
Pillow==9.1.0
pycodestyle==2.8.0
pytz==2022.1
qrcode==7.3.1
sqlparse==0.4.2
toml==0.10.2
tzdata==2022.1
